app/requirements.txt
black
Django==2.2.10
django-cors-headers
djangorestframework
drf-yasg==1.20
drf-yasg[validation]
djoser==2.0
flake8
isort
Pillow
psycopg2-binary
pytest-cov
pytest-django
pytest-xdist
pytest